Fixed Instance.__contains__.
Instance.__contains__ had a bug where it was using
Skeleton.node_to_index(node) in self._points. Skeleton.node_to_index
returns the index number of the node in the node list. self._points is
now keyed by node object so this was always returning false. Not sure
if this fixes everything with PAFs but it seems to be generating
something now.