Training now takes points instead of confmaps/pafs.
The points are passed as a list (each frame) of lists (each inst)
of point_arrays. Training also needs to be given a skeleton object
(we'll need to adjust this later to support multiple skeletons).
We can then run the Augmenter on the points and generate confmaps
or pafs after that from the augmented points. (Note we also have to
generate confmaps/pafs for the validation data.) Whether we gen
confmaps or pafs is determined by the output_type param.