bug fix to how we structure Instance/PredictedInstance from json
the cattr hook was checking the first item in the list for each frame
and then using that type for the whole list, so if first had score
then they all loaded as PredictedInstances. now we structure each
item in the list individually, checking its appropriate type as we go.