Defer loading MediaVideo reader until needed
This allows us to load labels object without loading the videos,
and if we don't need the videos, then it's fine if they aren't
available (i.e., the paths are wrong).
We also no longer raise an error if the path can't be found
(until we actually need the video).