Attempt to fix bugs from active learning results
The instances returned from active learning had nodes or skeletons
that weren't matching those already in the project, and this was
messing up lots of things. Now we merge video by video, and then
check nodes by name when creating instances from predicted instances
in the gui. This seems to fix issues 148 and 152.