Hash skeleton by id, remove __eq__ override
We were hashing skeletons by name, and __eq__ was checking for
same name and match(). This meant that serializing/deserializing
a skeleton would result in something == to the original so we'd
only save a single skeleton in the labels project (even if both
skeletons were used by instances).

Of course we'd usually want to unify the instances so they all
use the same skeleton (assuming the skeletons do match), but
the skeleton object shouldn't make things break if you don't unify.