Add unify arg to Labels.extend_from(), new tests.
We use unify if we want objects in the new frames to be replaced
with corresponding objects already in the labels.

Note that if we're adding two frames that could be unified with
each other, they won't be unified unless they can both be unified
with data already in the labels to which they're being added.