Suggestions grouped by video, interleave sortable.
The suggestions data table can now sort by group or by group
interleaved, and the order in the table can be saved.
Also fixed the display issue with the table not redrawing correctly
when you used one of the buttons in the same layout as the table
(this was fixed by calling processEvents before handling the
button action).