Write tracks as tuples to save space.
The hdf5 attr (where we save json for everything other than the
instance points) has a 64k limit, and we're going over that limit
when there are a ton of tracks. (Of course this isn't great anyway,
but we still shouldn't die when trying to write such a file.)
The json we're saving is smaller if we save tracks as tuples, rather
than dict (with key names) for each track.