Save videos/tracks/suggestions in hdf5 datasets.
We were saving them (with all other json-formatted metadata) in
an attribute on the metadata dataset. But attributes have 64k
limit, which was a problem when we had lots of tracks, videos, or
suggestions.

The new functionality isn't used when *appending* to an hdf5 file
(although I don't think we currently use this anywhere except
tests). If the data can't be found in new datasets, it wil be
expected in the metadata attribute (as before) for backwards
compatibility.

Fixes issues #243.