Add info.align module for making "mean" instance.
Module has functions to align a bunch of instances (and find the
best nodes to use for alignment), and then calculate a "mean"
instance from the aligned instances.

In the gui we use this as a template when the user creates a new
instance or converts a predicted instance w/ missing points to an
editable instance.