Support mixed-video backend for training packages.
When you export a training package with select video frames, we
now also save the information about the source video. If we then
try to get frames that weren't in the training package, we'll also
try to load them from the source video if possible.

This can be disabled by setting enable_source_video to False for
the HDF5Video backend.

Resolves issue #270.