Run tracking/create instances frame by frame.
We had been running inference for all the frames and then running
tracking and creating the instances for the entire chunk of frames.
This didn't work well when using flow shift since we were keeping
the full images for all frames.

Now we use an interator to group examples that are from the same
frame, and make the instances/run tracking while looping over this.