Main and develop should be at a singularity (#821)
Main is listed as being 56 commits behind develop, but in reality 0 files are changed between main and develop at this point in time - the git history is just a little scrambled. This commit aims to rectify this scrambled history which will be beneficial at the next release/merge into main.